requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
billiard==4.2.1
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kombu==5.5.4
lxml==5.4.0
MarkupSafe==3.0.2
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prisma==0.15.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.0
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
tomlkit==0.13.2
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
aio_pika==8.3.0
